Refactor 40percentclub/ut47
