+++ keyboards/40percentclub/ut47/keyboard.json
@@ -22,6 +22,15 @@
       "resync": true
     }
   },
+  "diode_direction": "COL2ROW",
+  "matrix_pins": {
+    "cols": ["D7", "E6", "B4", "B5", "B6", "B2", "B3", "B1", "F7", "F6", "F5", "F4"],
+    "rows": ["D1", "D0", "D4", "C6"]
+  },
+  "keycodes": [
+    {"key": "LED_TOG"},
+    {"key": "LED_CHG"}
+  ],
   "layouts": {
     "LAYOUT": {
       "layout": [
