+++ keyboards/40percentclub/ut47/keymaps/default/keymap.c
@@ -1,30 +1,9 @@
-/* Copyright 2018 Carlos Filoteo
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
+// Copyright 2018 Carlos Filoteo
+// SPDX-License-Identifier: GPL-2.0-or-later
 #include QMK_KEYBOARD_H
-#ifdef LED_ENABLE
-  #include "uart.h"
-#endif
 
 #define LT3_TAB LT(3, KC_TAB)
 #define MT_RSFT_ENT MT(MOD_RSFT, KC_ENT)
-
-enum custom_keycodes {
-    LED_TOG = SAFE_RANGE,
-    LED_CHG
-};
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 
@@ -104,32 +83,3 @@ LAYOUT( /* Tab */
   _______, _______, _______, _______, _______,     _______,      _______, MS_LEFT, MS_DOWN, MS_UP,   MS_RGHT
 ),
 };
-
-//LED keymap functions
- #ifdef LED_ENABLE
-void led_chmode(void) {
-  uart_write(0x65);
-}
-
-void led_toggle(void) {
-  uart_write(0x64);
-}
-
-bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-  if (record->event.pressed) {
-    switch(keycode) {
-      case LED_TOG:
-        #ifdef LED_ENABLE
-        led_toggle();
-        #endif
-        return false;
-      case LED_CHG:
-        #ifdef LED_ENABLE
-        led_chmode();
-        #endif
-        return false;
-    }
-  }
-  return true;
-};
-#endif
