+++ keyboards/40percentclub/ut47/post_rules.mk
@@ -1,4 +1,5 @@
 ifeq ($(strip $(LED_ENABLE)), yes)
-  OPT_DEFS += -DLED_ENABLE
-  SRC += led.c
+    OPT_DEFS += -DLED_ENABLE
+    SRC += led_controls.c
+    UART_DRIVER_REQUIRED = yes
 endif
