--- keyboards/40percentclub/ut47/rules.mk
@@ -1,3 +1,0 @@
-CUSTOM_MATRIX = yes
-SRC += matrix.c
-UART_DRIVER_REQUIRED = yes
